postgres: support non-alphanumeric usernames
Currently postgres fails to create a database because my username
contains a `.`. Wrapping the database name in `"` will resolve the
problem.

I've rewritten the line to use HEREDOC so that we do not have to mess
with escaping (from Nix) within escaping (from Bash) within escaping
(from psql). Now it only escapes using Nix and psql.