devenv: avoid rebuild when unrelated sources change
Currently whenever the examples are changed, devenv will need to be
rebuild. This isn't needed, because the package isn't relying on the
examples explicity.

Using `builtins.path` was just too much of a hassle (paths are absolute,
directory path needs to be match along with sub-files/directories).
Regex seemed easier.

Regex in Nix uses extended Posix, which is again a bit funky, but maybe
less so than doing matching by hand.